(feature): Add the concept of 'orphaned' events
Event Instances are associated with event 'periods'.
When an event period is deleted, or changes dates, then this may lead to
some event instances no longer falling within the date range
defined by any period.

The approach is to allow the admin to optionally delete these, or to
leave them (e.g. if a new event period is about to be created which they
will fall within)