(delete): Don't handle existing EventInstances
The code was trying to account for the case where we're generating an
event instance here one already exists and making sure we return an
array representing what we did for each date (returned the existing one
or created a new one).

This is probably YAGNI - certainly isn't used currently, because it's
only applicable to the 'create' case where we're creating an event from
scratch, so nothing will already exist.