(refactor): Adding dates = adding event instances
Context: we've redefined what an event period/generator is for one-off
and intermittent events (no longer the 'thing which generates that one
instance' - now 'the concept of a period of time during which the event
is basically the same').

So now adding when adding dates we shouldn't create new event periods
(assume that an open-ended one was created when the event was) - we just
need to create the instances directly.