(feature): Generate one-offs, even if in the past
This reverses an earlier decision. The thinking now is that even if an
event date was in the past, we should be allowed to create it - e.g. for
the purposes of calculating the next date of a monthly repeating event.

Note that the event period spec covers the 'generate' method, which is
now better covered in spec/features/generate_event_instances_spec.rb
- for now we'll keep hold of that part of the spec, but mostly defer to
the integration spec.