(feature): Start and end Period - use datepicker
Use Pickadate date pickers on the start and end event period forms.

This isn't great because it's not tested: the tests hitting these pages
don't run using javascript, so as far as Capybara is concerned, this is
just a text field. That makes it possible for these tests to run (there
doesn't seem to be any easy way of interacting with pickadate through
capybara), but if we ever turn on js for specs interacting with these
forms then those specs will break because it won't be possible to
directly fill in a date in these fields.