(fix): EventInstance dates aren't restricted
Event instances had validations preventing them from being created too
far in the past or the future - this is intended to prevent obvious
mistakes.

This causes problems when importing events from the old database: there
are dates further in the past than this range allows, so the import
failed.

The point at which we actually care about setting these dates is when
creating a new event. There are validations on the event form object
which implement the same date range, so this restriction can be removed
from the model.