(feature): use dd/mm/yyyy
Switch to using dd/mm/yyyy as a default instead of yyyy-mm-dd.

Use locales:

Pros: date format lives in a locale where it belongs
Cons: have to explicitly format each date with `l`

The test approach is to either use explicit dates, or to use `I18n.l` in
the specs, which is questionable, but maybe OK since the locales can be
considered config.

There are a couple of alternate approaches to this. The main one would be:

    # config/initializers/time_formats.rb
    Date::DATE_FORMATS[:default]="%Y/%m/%d"

Pros: don't need to explicitly format each date
Cons: the behaviour is a bit too implicit (hard to work out where it's
set) and any overriding becomes an ugly `my_date.to_s(:my_format)`