(feature) end a repeating period
Adapted from the 'scheduling a break' feature. Might go back and revisit
that later (specifying a date range that a thing won't be running), but
for now this feels like the feature that makes most sense.

- Stop using edit/update for this: create a new controller responsible
for ending event periods which are repeating
- Nothing really 'ends' - at least not through event periods. Instead
things which were weekly can become intermittent, so that if a new event
instance happens (a one-off instance of something that used to be
weekly) then that is understood to happen during an intermittent period.