NOTES:
it's really hard to come up with a good data model!

Removing end date breaks the event generation stuff

I think the current approach is just wrong:
"go through each event period and decide whether to generate instances
for it"

Instead a better approach might be:
"go through each event, and for the next 4 weeks decide:
a) should an event be auto-generated for this date?
b) what data should be used for that date?"

This would require adding a start date onto event seeds as well, so that
event seeds become a sort of immutable list of the changes which were
made to the event data.

It also removes the idea that events can be repeating with a frequency
other than 0 or 1, but maybe that was never really a thing anyway...

There's a reversal here: instead of working out what the dates should be
in the next month, go through each of the possibilities (because at most
it's weekly)

There are two cases when we need to know what data (event seed) to use:
when adding a one-off instance, and when generating instances for the
next month. The proposal here is that these both be reduced to the
problem of checking which event seed applies for a specific date.